Checkpoints
Added a start, checkpoint and a finish.

You are forced to go from the start to the checkpoint. When you passed
the (last) checkpoint you are allowed to finish.
Multiple checkpoints are possible, because you should assign the
previous checkpoint.

What doesn't work:
finish doesn't stop the game.
checkpoints can't be reset so multiple laps are not possible yet