Alot of adjustments
Added a GameManager to keep track of the state of a game.
Some other files needed to be adjusted to use the GameManager.
NetworkManager is enabled again but multiplayer doesn't really work
because the PlayerController is still broken.
PlayerController only works for moving a block not a car, atleast for
now.
I should fix the multiplayer soon again.

I made the car and wheels this time heavier so the car is effected a
little bit less by the cones.

NetworkManager GUI should change but wondering about StartServer being a
public function.