Best roundTime
Code keeps track of the best roundtime (per session)

Also added acceleration code to the CarController.
It works with using a max and min troque. Problem here is, there is no
max speed. This should be added.
Antirollbars should also be added so the car takes corners a little bit
better.